back/requirements.txt
asgiref==3.4.1
bleach==4.1.0
certifi==2021.10.8
charset-normalizer==2.0.7
coreapi==2.3.3
coreschema==0.0.4
Django==3.2.8
django-appconf==1.0.5
django-cors-headers==3.10.0
django-extensions==3.1.3
django-imagekit==4.1.0
django-summernote==0.8.20.0
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
drf-yasg==1.20.0
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.2
MarkupSafe==2.0.1
packaging==21.0
pilkit==2.0
Pillow==8.4.0
PyJWT==1.7.1
pyparsing==3.0.3
pytz==2021.3
requests==2.26.0
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.7
webencodings==0.5.1
